+++ src/index.tsx
@@ -4,7 +4,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
-import store from "./redux/DiaryCardSlice/Store";
+import store from "./redux/Store";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
